Split extended directory
