+++ test/extended-priv/mco_password.go
@@ -14,8 +14,8 @@ import (
 	expect "github.com/google/goexpect"
 	g "github.com/onsi/ginkgo/v2"
 	o "github.com/onsi/gomega"
-	exutil "github.com/openshift/machine-config-operator/test/extended/util"
-	logger "github.com/openshift/machine-config-operator/test/extended/util/logext"
+	exutil "github.com/openshift/machine-config-operator/test/extended-priv/util"
+	logger "github.com/openshift/machine-config-operator/test/extended-priv/util/logext"
 	"golang.org/x/crypto/ssh"
 	e2e "k8s.io/kubernetes/test/e2e/framework"
 )
