+++ test/extended-priv/secret.go
@@ -6,8 +6,8 @@ import (
 	"path/filepath"
 	"time"
 
-	exutil "github.com/openshift/machine-config-operator/test/extended/util"
-	logger "github.com/openshift/machine-config-operator/test/extended/util/logext"
+	exutil "github.com/openshift/machine-config-operator/test/extended-priv/util"
+	logger "github.com/openshift/machine-config-operator/test/extended-priv/util/logext"
 	e2e "k8s.io/kubernetes/test/e2e/framework"
 )
