+++ test/extended-priv/skopeo_client.go
@@ -7,7 +7,7 @@ import (
 	"os/exec"
 	"strings"
 
-	logger "github.com/openshift/machine-config-operator/test/extended/util/logext"
+	logger "github.com/openshift/machine-config-operator/test/extended-priv/util/logext"
 	e2e "k8s.io/kubernetes/test/e2e/framework"
 )
