+++ test/extended-priv/testdata/files/custom-machine-config-pool.yaml
