+++ test/extended-priv/testdata/files/generic-machine-config-template.yml
