+++ test/extended-priv/testdata/files/generic-machine-os-config-nobaseimagepull.yaml
