+++ test/extended-priv/testdata/files/generic-machine-os-config.yaml
