+++ test/extended-priv/testdata/files/machineconfigurations/managedbootimages-all.yaml
