+++ test/extended-priv/testdata/files/machineconfigurations/managedbootimages-empty.yaml
