+++ test/extended-priv/testdata/files/machineconfigurations/managedbootimages-none.yaml
