+++ test/extended-priv/testdata/files/machineconfigurations/managedbootimages-partial.yaml
