+++ test/extended-priv/types.go
