+++ test/extended-priv/util/client.go
