+++ test/extended-priv/util/resource_op.go
