+++ test/extended/boot_image_update_agnostic.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	machineclient "github.com/openshift/client-go/machine/clientset/versioned"
-	exutil "github.com/openshift/machine-config-operator/test/extended/util"
+	exutil "github.com/openshift/machine-config-operator/test/extended-priv/util"
 
 	o "github.com/onsi/gomega"
